+++ lib/src/menu/menu.dart
@@ -1,5 +1,6 @@
 /// Provides the [Menu] class.
-import '../command.dart';
+import 'package:dart_sdl/dart_sdl.dart';
+
 import '../game.dart';
 import '../json/asset_reference.dart';
 import '../json/message.dart';
@@ -26,8 +27,23 @@ class Menu extends Level {
       {required Game game,
       required this.title,
       List<MenuItem>? items,
-      List<Ambiance>? ambiances})
+
+      /// The position within the menu.
+      ///
+      /// If this value is `null`, then the title will be focused.
+      int? position,
+      this.onCancel,
+      List<Ambiance>? ambiances,
+      this.upScanCode = ScanCode.SCANCODE_UP,
+      this.upButton = GameControllerButton.dpadUp,
+      this.downScanCode = ScanCode.SCANCODE_DOWN,
+      this.downButton = GameControllerButton.dpadDown,
+      this.activateScanCode = ScanCode.SCANCODE_RIGHT,
+      this.activateButton = GameControllerButton.dpadRight,
+      this.cancelScanCode = ScanCode.SCANCODE_LEFT,
+      this.cancelButton = GameControllerButton.dpadLeft})
       : menuItems = items ?? [],
+        _position = position,
         super(game, ambiances: ambiances ?? []);
 
   /// The title of this menu.
@@ -36,34 +52,41 @@ class Menu extends Level {
   /// The menu items contained by this menu.
   final List<MenuItem> menuItems;
 
+  /// The function that will be called by [cancel].
+  ///
+  /// If this value is `null`, nothing will happen when the [cancel] method is
+  /// called.
+  final void Function()? onCancel;
+
+  /// The scancode that will call the [up] method.
+  final ScanCode upScanCode;
+
+  /// The scancode that will call the [down] method.
+  final ScanCode downScanCode;
+
+  /// The scancode that will call the [activate] method.
+  final ScanCode activateScanCode;
+
+  /// The scancode that will call the [cancel] method.
+  final ScanCode cancelScanCode;
+
+  /// The button that will call the [up] method.
+  final GameControllerButton upButton;
+
+  /// The button that will call the [down] method.
+  final GameControllerButton downButton;
+
+  /// The button that will call the [activate] method.
+  final GameControllerButton activateButton;
+
+  /// The button that will call the [cancel] method.
+  final GameControllerButton cancelButton;
+
   //// The current position in this menu.
   int? _position;
 
   /// The last sound played by this menu.
   PlaySound? oldSound;
-
-  /// Register default commands.
-  ///
-  /// If any of the arguments are `null`, then that command will not be
-  /// registered.
-  void registerCommands(
-      {String? upCommandName,
-      String? downCommandName,
-      String? activateCommandName,
-      String? cancelCommandName}) {
-    if (upCommandName != null) {
-      registerCommand(upCommandName, Command(onStart: up));
-    }
-    if (downCommandName != null) {
-      registerCommand(downCommandName, Command(onStart: down));
-    }
-    if (activateCommandName != null) {
-      registerCommand(activateCommandName, Command(onStart: activate));
-    }
-    if (cancelCommandName != null) {
-      registerCommand(cancelCommandName, Command(onStart: cancel));
-    }
-  }
 
   /// Show the current item.
   void showCurrentItem() {
@@ -108,7 +131,12 @@ class Menu extends Level {
   }
 
   /// What happens when this menu is cancelled.
-  void cancel() {}
+  void cancel() {
+    final f = onCancel;
+    if (f != null) {
+      f();
+    }
+  }
 
   /// Get the currently focussed menu item.
   MenuItem? get currentMenuItem {
@@ -162,5 +190,37 @@ class Menu extends Level {
         MenuItem(Message(sound: selectSound, text: text), menuItemLabel);
     menuItems.add(item);
     return item;
+  }
+
+  /// Handle SDL events.
+  @override
+  void handleSdlEvent(Event event) {
+    if (event is KeyboardEvent &&
+        event.repeat == false &&
+        event.state == PressedState.pressed &&
+        event.key.modifiers.isEmpty) {
+      final scanCode = event.key.scancode;
+      if (scanCode == upScanCode) {
+        up();
+      } else if (scanCode == downScanCode) {
+        down();
+      } else if (scanCode == activateScanCode) {
+        activate();
+      } else if (scanCode == cancelScanCode) {
+        cancel();
+      }
+    } else if (event is ControllerButtonEvent &&
+        event.state == PressedState.pressed) {
+      final button = event.button;
+      if (button == upButton) {
+        up();
+      } else if (button == downButton) {
+        down();
+      } else if (button == activateButton) {
+        activate();
+      } else if (button == cancelButton) {
+        cancel();
+      }
+    }
   }
 }